Update get_data.py
Remove a lot of inefficiencies in the code when there are multiple tags. Rather than getting data one by one based on the config file, the get_data file will listen to all available ruuvitag updates, take the average value for each tag and it's respective value before sending it.